+++ database/seeders/DatabaseSeeder.php
@@ -14,6 +14,7 @@ class DatabaseSeeder extends Seeder
         $this->call([
             RolePermissionSeeder::class,
             DepartmentSeeder::class,
+            DocumentTypeSeeder::class,
             DivisionSeeder::class,
             UserSeeder::class,
             SettingSeeder::class,